CertificateTest: fix error-prone code in test
There was a comparison of a byte against the value 200, which would
always fail. From the code below it's evident that the comparison should
be against the value 200 casted as a byte.

Test: vogar CertificateTest

+++ luni/src/test/java/tests/security/cert/CertificateTest.java
@@ -174,7 +174,7 @@ public class CertificateTest extends TestCase {
         byte[] actualEncoding = actual.getEncoded();
         assertTrue(Arrays.equals(expectedEncoding, actualEncoding));
 
-        assertFalse(expectedEncoding[4] == 200);
+        assertFalse(expectedEncoding[4] == (byte) 200);
         expectedEncoding[4] = (byte) 200;
         try {
             cf.generateCertificate(new ByteArrayInputStream(expectedEncoding));
